+++ src/main/java/baron/command/Command.java
@@ -1,3 +1,10 @@
+package baron.command;
+
+import baron.Storage;
+import baron.Ui;
+import baron.exception.BaronException;
+import baron.task.Task;
+
 import java.util.ArrayList;
 
 public abstract class Command {
