+++ src/main/java/baron/command/EventCommand.java
@@ -1,3 +1,10 @@
+package baron.command;
+
+import baron.Storage;
+import baron.Ui;
+import baron.task.EventTask;
+import baron.task.Task;
+
 import java.time.LocalDateTime;
 import java.util.ArrayList;
