+++ src/main/java/baron/exception/BaronException.java
@@ -1,2 +1,4 @@
+package baron.exception;
+
 public class BaronException extends Exception {
 }
