+++ src/main/java/baron/exception/InvalidTaskIndexException.java
@@ -1,2 +1,4 @@
+package baron.exception;
+
 public class InvalidTaskIndexException extends BaronException {
 }
