+++ src/main/java/baron/exception/ReservedCharacterException.java
@@ -1,2 +1,4 @@
+package baron.exception;
+
 public class ReservedCharacterException extends BaronException {
 }
