+++ src/main/java/baron/exception/WrongUsageException.java
@@ -1,2 +1,4 @@
+package baron.exception;
+
 public class WrongUsageException extends BaronException {
 }
