+++ src/main/java/baron/task/Task.java
@@ -1,3 +1,5 @@
+package baron.task;
+
 public abstract class Task {
     private final String taskName;
     private boolean isDone;
