+++ src/main/java/baron/task/ToDoTask.java
@@ -1,3 +1,5 @@
+package baron.task;
+
 public class ToDoTask extends Task {
     public ToDoTask(String taskName) {
         this(false, taskName);
