+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Baron.
+Hello! I'm baron.baron.
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
